pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
openai = "^0.27.9"
rich = "13.5.2"
loguru = "^0.7.0"
google-cloud-dialogflow = "2.23.3"
chatterbot-voice = "^0.1.5"
bit = "^0.8.0"
pillow = "^10.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"